+++ acceptance-test/src/test/java/br/com/helpdev/atdd/DefaultContainerStarterTest.java
@@ -4,6 +4,7 @@ import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMoc
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import io.restassured.RestAssured;
+import org.junit.jupiter.api.AfterEach;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.GenericContainer;
@@ -14,52 +15,52 @@ import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.lifecycle.Startable;
 import org.testcontainers.utility.MountableFile;
 
-abstract class AbstractContainerBaseTest {
+abstract class DefaultContainerStarterTest {
 
   private static final GenericContainer<?> APP;
   private static final GenericContainer<?> FLYWAY;
   private static final GenericContainer<?> MYSQL_CONTAINER;
+  private static final Network NETWORK = Network.newNetwork();
   protected static final WireMockServer MOCK_SERVER;
 
+  /* Containers are initialized in static block to create only once in test execution  */
   static {
-    final var network = Network.newNetwork();
-
     MOCK_SERVER = new WireMockServer(wireMockConfig().dynamicPort());
     MOCK_SERVER.start();
     exposeHostMachinePortToContainersForApiIntegrations();
 
-    MYSQL_CONTAINER = buildMySqlContainer(network);
+    MYSQL_CONTAINER = buildMySqlContainer();
     MYSQL_CONTAINER.start();
 
-    FLYWAY = buildFlywayContainer(network, MYSQL_CONTAINER);
+    FLYWAY = buildFlywayContainer(MYSQL_CONTAINER);
     FLYWAY.start();
 
-    APP = buildAppContainer(network, MYSQL_CONTAINER, FLYWAY);
+    APP = buildAppContainer(MYSQL_CONTAINER, FLYWAY);
     APP.start();
 
     initRestAssured();
   }
 
-  private static GenericContainer<?> buildMySqlContainer(final Network network) {
+  private static GenericContainer<?> buildMySqlContainer() {
     return new MySQLContainer<>("mysql:5.7.22")
-        .withNetwork(network)
+        .withNetwork(NETWORK)
         .withNetworkAliases("testdb");
   }
 
-  private static GenericContainer<?> buildFlywayContainer(final Network network, final Startable... dependsOn) {
+  private static GenericContainer<?> buildFlywayContainer(final Startable... dependsOn) {
     return new GenericContainer<>("flyway/flyway")
         .dependsOn(dependsOn)
-        .withNetwork(network)
+        .withNetwork(NETWORK)
         .withCopyFileToContainer(MountableFile.forHostPath("../resources/flyway/db"), "/flyway/sql")
         .withCommand("-url=jdbc:mysql://testdb?useSSL=false -schemas=test -user=test -password=test -connectRetries=60 migrate")
         .waitingFor(Wait.forLogMessage("(?s).*No migration necessary(?s).*|(?s).*Successfully applied(?s).*", 1))
         .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("FLYWAY")));
   }
 
-  private static GenericContainer<?> buildAppContainer(final Network network, final Startable... dependsOn) {
+  private static GenericContainer<?> buildAppContainer(final Startable... dependsOn) {
     return new GenericContainer<>("app-test:integration")
         .dependsOn(dependsOn)
-        .withNetwork(network)
+        .withNetwork(NETWORK)
         .withEnv("RANDOM_DATA_API_URL", "http://host.testcontainers.internal:" + MOCK_SERVER.port())
         .withEnv("MYSQL_USER", "test")
         .withEnv("MYSQL_PASSWORD", "test")
@@ -78,6 +79,12 @@ abstract class AbstractContainerBaseTest {
 
   private static void exposeHostMachinePortToContainersForApiIntegrations() {
     Testcontainers.exposeHostPorts(MOCK_SERVER.port());
+  }
+
+  @AfterEach
+  void tearDown() {
+    MOCK_SERVER.resetAll();
+    /* add here others resets needed after each test */
   }
 
 }
